% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{DEGage}
\alias{DEGage}
\title{DEGage}
\usage{
DEGage(
  counts,
  group,
  perm.preprocess = TRUE,
  gene.filter.threshold = 1,
  nperms = 2000
)
}
\arguments{
\item{group}{A factor which assigns conditions to cells as in Deseq2.}

\item{perm.preprocess}{A logical indicating whether or not to perform the permutation prefiltering step.}

\item{gene.filter.threshold}{A value between 0-1 which represents the maximum proportion of zeros a gene can have before being filtered out.}

\item{nperms}{An integer greater than 0 that indicates how many permutations will be carried out during the permutation test}

\item{input}{A dataframe with cells as columns and genes as rows.}
}
\description{
Tests for pairwise differential expression between two count distributions
}
