% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{DEGage_pdf}
\alias{DEGage_pdf}
\title{DEGage_pdf}
\usage{
DEGage_pdf(r1, p1, r2, p2, k)
}
\arguments{
\item{r1}{r parameter for the first NB distribution}

\item{p1}{p parameter for the first NB distribution}

\item{r2}{r parameter for the second NB distrbution}

\item{p2}{p parameter for the second NB distribution}

\item{k}{Observed number of differences}
}
\description{
Calculates pvalues from regression parameters
}
