% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{gen_colnames}
\alias{gen_colnames}
\title{gene.colnames}
\usage{
gen_colnames(ncells, ngroup1)
}
\arguments{
\item{ncells}{the number of cells to be generated}

\item{ngroup1}{the number of cells in group one}
}
\description{
A helper function that generates a name for each cell for simulation. Names indicate which condition/group each cell belongs to
}
