% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{gen_rownames}
\alias{gen_rownames}
\title{gen_rownames}
\usage{
gen_rownames(ngenes, ndegs)
}
\arguments{
\item{ngenes}{the number of total genes}

\item{ndegs}{the number of desired differentially expressed genes}
}
\description{
A helper function that generates row names for simulation.
}
