% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{pdf_facilitator}
\alias{pdf_facilitator}
\title{pdf_facilitator}
\usage{
pdf_facilitator(df)
}
\arguments{
\item{df}{The data frame containing regression parameters for every gene}
}
\description{
A helper function that calls upon DEGage_pdf
}
