% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permtest}
\alias{permtest}
\title{permtest}
\usage{
permtest(genecount, group, levels, nperms)
}
\arguments{
\item{genecount}{An R Numeric vector containing counts for a single gene}

\item{group}{An R Numeric vector containing group info, typically two integers that correspond to groups}

\item{levels}{The levels aspect of an R factor for the above group parameter}

\item{nperms}{The number of permutations to perform}
}
\description{
Performs genewise permutation tests to calculate pvalues for prefiltering.
}
