% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{QC}
\alias{QC}
\title{QC}
\usage{
QC(counts, group, gene.filter.threshold, nperms)
}
\arguments{
\item{group}{A factor which assigns conditions to cells as in Deseq2.}

\item{gene.filter.threshold}{A value between 0-1 which represents the maximum proportion of zeros a gene can have before being filtered out.}

\item{nperms}{An integer greater than 0 that indicates how many permutations will be carried out during the permutation test}

\item{input}{A dataframe with cells as columns and genes as rows.}
}
\description{
Checks user input to ensure DEGage can run properly
}
