% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{zeroinfl_model_fitting}
\alias{zeroinfl_model_fitting}
\title{zeroinfl_model_fitting}
\usage{
zeroinfl_model_fitting(counts, group)
}
\arguments{
\item{counts}{A sliced dataframe that contains the counts for an individual gene.}

\item{group}{A factor that contains grouping information for the counts}
}
\description{
A helper function that performs genewise zero-inflated negative binomial regression
}
